limit font-size + refactor
